ui: fix flaky test for "query" sagas
Initially, test logic was defined in a way that query
should be auto refreshed every 50ms and `silentRun` limited
saga execution time to 250ms only, so it was enough to queue
tasks and auto-refresh it maximum 5 times due to timing limitation
but not because of business logic.
Sometimes this test failed when task was executed 6 times and it
is valid expected result, possibly due to asynchronous nature of
sagas and how it waits for next actions.

Now, this test modified to have small `refreshInterval` value
and it allows to run all tasks in queue before timeout.

Expected value is changed to 6 because:
- `queryCalled` is incremented 3 times when `autoRefresh` action
is dispatched 3 times in a row manually
- and then it is called 3 times for every `stopAutoRefresh` action
respectively

Release note: None